requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.2
Django==3.2.6
django-allauth==0.45.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-uuidfield==0.5.0
djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.1
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
filelock==3.0.12
flex==6.14.1
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonpointer==2.1
Markdown==3.3.4
MarkupSafe==2.0.1
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==21.0
platformdirs==2.2.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rfc3987==1.3.8
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
simplejson==3.17.3
six==1.15.0
sqlparse==0.4.1
strict-rfc3339==0.7
uritemplate==3.0.1
urllib3==1.26.6
validate-email==1.3
virtualenv==20.7.0
